Fix issue with creating novels
Fixed an issue in which creating novels in the back end was not working. Dates were not being inserted correctly. Also fixed display issue with some of the auth forms. Added list of novesl to top of Novels page to make it easier to susbscribe (don’t have to scroll all the way down the page to see the list).